requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
greenlet>=3.0.0
alembic>=1.13.0
pyepics>=3.5.2
python-multipart>=0.0.6
httpx>=0.26.0

# Dev dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
ruff>=0.1.0
mypy>=1.8.0
pre-commit>=3.6.0

# Documentation 
mkdocs
mkdocs-material
mkdocstrings
mkdocstrings-python
